PEVC-dealbox-agent/requirements.txt
firecrawl-py
groq
python-dotenv
pydantic